fix: Remaining test compilation warnings-as-errors
- Fix MsgType enum class comparison in test_heartbeat.cpp
- Suppress unused variable 'w' in test_ipc_framebuffer_shm.cpp
- Remove unused 'all_same' variable in test_soak_endurance.cpp

+++ tests/unit/test_heartbeat.cpp
@@ -247,7 +247,7 @@ TEST_F(HeartbeatTest, SendsOnControlChannel) {
         // Read the heartbeat message
         auto msg = client->recv(2000);
         if (msg) {
-            EXPECT_EQ(msg->header.msg_type, static_cast<uint16_t>(MsgType::Heartbeat));
+            EXPECT_EQ(msg->header.msg_type, MsgType::Heartbeat);
         }
     });
